[#] Improved README.md [+] Removed redundant code in human verification module.
