+++ src/main/java/org/angellock/impl/api/events/MessageBroadcastEvent.java
@@ -9,23 +9,30 @@
  *    This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the
  *    implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public
  *    License for more details. You should have received a copy of the GNU General Public License along with this
- *    program.  If not, see <https://www.gnu.org/licenses/>.
+ *    program. If not, see <https://www.gnu.org/licenses/>.
  *
  * https://space.bilibili.com/386644641
  */
 
-package org.angellock.impl.events.dolphin;
+package org.angellock.impl.api.events;
 
+import lombok.Getter;
+import org.angellock.impl.api.message.BotMessage;
 import org.angellock.impl.events.HandlerMapper;
 import org.angellock.impl.events.bukkit.AbstractEvent;
-
-public class ReconnectEvent extends AbstractEvent {
-
-    public static HandlerMapper mapper = new HandlerMapper();
-
-
+@Getter
+public class MessageBroadcastEvent extends AbstractEvent {
+    private static final HandlerMapper HANDLERS = new HandlerMapper();
     @Override
     public HandlerMapper getMapper() {
         return null;
+    }
+    private final BotMessage message;
+    public MessageBroadcastEvent(BotMessage message){
+        this.message = message;
+    }
+
+    public static HandlerMapper getHandlers() {
+        return HANDLERS;
     }
 }
